Medium: Tools: hb_report: options -C and -D are obsoleted
-D is used to disable editing description. Since almost nobody
is using this option and most often the report is an attachment
to either bugzilla or mail message, hb_report now defaults to
*not* editing the description.

-C is used to disable removing directories which were used to
create per-node and cummulative reports. It turned out that this
is unnecessary producing trash on nodes and now hb_report
defaults to removing the directories. In future, there will be a
debugging option which would keep the directories. It seems like
otherwise nobody really wants these directories.