Medium: cibconfig: refactor elements edit operation
A major rewrite of the edit operation. It is relegated to the
CibFactory class. Fixes some corner cases which used to fail
before (in particular, moving an element from one container to
another).

There is now also a rollback functionality in CibFactory. It
prevents half edited configurations which could happen before.

Rollback is not yet exposed to the user.

The moved and invalid CibObject attributes are obsolete.