Dev: userprefs: refactor
Every user preference is an instance of a class which corresponds
to the preference type (boolean, string, etc). Makes user
preferences streamlined and much easier to add new.

The line between user options (as specified on the command line)
and preferences is thin. Some options are handled as preferences.
Though that probably shouldn't matter. Perhaps more options
should move to preferences (but not to be saved to the rc file).

Thanks to Kristoffer Grönlund for help with python classes
modifications (don't use self.__dict__ in self.__getattr__).