utils: Don't crash on missing reply to y/n question
If no reply is given (only enter is hit), the question
seemed to imply that y would be chosen. This was not the
intention with the code, but the code for that case was
also incorrect. This commit fixes both, so that the question
does not imply an implicit response, and the question repeats
if no answer is given.