high: parse: Try to retain ordering if possible (bnc#880371)
The CIB is generally not order-sensitive, except for one thing:
in containers, primitive elements are internally order-sensitive.

This is complicated by the fact that containers also contain other tags
(attribute lists) which are not order-sensitive.

This commit tries to retain the ordering of those tags like pacemaker
tends to order them, while being order-sensitive with respect to
primitive tags.

XML sucks.