high: parse: Implicit initial parameter list
This adds support for a shorthand syntax when creating
primitives and nodes: Instead of explicitly writing

primitive foo Foo params foo-param=1

this is now accepted as equivalent:

primitive foo Foo foo-param=1

The XML->CIB parser is still the same, so when inspecting
the created CLI syntax, params will have been injected.
However, this should make the creation of new primitives
a bit cleaner.