high: utils: Locate binaries across sudo boundary (bnc#912483)
When the core.user option is set, the PATH variable is modified to
include the pacemaker lib dir, but when invoking a child process
via sudo -E, the modification is lost. Most likely this is
related to the security policy.

The workaround is to make sure that the absolute path to the
binary is used.

Conflicts:
	modules/utils.py