medium: command: Fuzzy match command names
Adds the feature of fuzzy matching command names.
The fuzzy match works in three layers:

If the command name is an exact match, use that.
If the command name is a unique prefix string, use that.
If the command name is a prefix substring, use that.

Prefix substring: Used name starts with the same letter and
contains all the letters of a command in the given order.

For example, "cfg" is a prefix substring of configure.
"ss" is a prefix substring of status.
"st" is a prefix of status.