dev: Fix installation issue for Debian (#59)
On Debian, python packages should install to the directory
dist-packages and not site-packages. However, if we pass
--prefix= to setup.py, the modification for Debian is not applied
and we end up installing in the wrong location. This checks
if the prefix is equal to /usr, and if so we avoid passing --prefix=.

This should allow the installation to work on Debian systems as well.