#### Fix build on Debian ***
- Updated AM_CONDITIONAL defining `PREFIX_IS_USR`, to define
`UNAME_IS_DEBIAN` instead, by checking the version of the LSB
specification against which the host distribution is compliant, in order
to check whether we are building for Debian or not.
- Updated `if`/`fi` conditional check in `Makefile.am`, so that it does
a check on - `UNAME_IS_DEBIAN` instead of `PREFIX_IS_USR`.
- Added simple ignores in .gitignore for Eclipse and PyDev Eclipse
Plugin based files, as well as GEdit files.

Change-Id: Ie704dd6dcd752f44e2adeff93fb32f03d56c8cb0
Signed-off-by: Richard B Winters <rik@mmogp.com>