dev: Convert to relative imports
Implicitly relative imports are no longer allowed in python 3, so
as a first step towards making crmsh run on python 3, convert all
such imports to relative imports.

There are some complications in getting the crm script and the unit
tests to correctly load and run crmsh as a module in the development
environment.