build: Add update-data-manifest.sh to generate datadir file list
Revised automake to use data-manifest for all files to install
into /usr/share/crmsh.

data-manifest is generated by running update-data-manifest.sh. The
script uses git ls-files to avoid including garbage files excluded
by .gitignore.

The data-manifest file has to be updated by rerunning the script
whenever a file is added or removed.

This commit also adds the unittests to the list of installed files.
That means that they will be included in the -test package.