medium: parse: Add support for node attribute as fencing topology target
Pacemaker's upstream master branch has a new feature that will be part
of the eventual 1.1.14 release.

Fencing topology is used when a node requires multiple fencing devices
(in combination or as fallbacks). Currently, topologies must be
specified by node name (or a regular expression matching node names).

The new feature allows topologies to specified by node attribute.

For example, imagine a data center where all devices in rack #1 use
fence devices apc01 and apc02, while all devices in rack #2 use fence
devices apc03 and apc04.

Previously, if node1 was in rack #1, you'd have to register a fencing
topology by its name, which at the CLI level would look like:

   fencing-topology node1: apc01,apc02

With the new feature, you could instead register a topology for all
hosts that have a node attribute "rack" whose value is "1":

   fencing-topology rack=1: apc01,apc02

You would assign that attribute to all nodes in that rack, e.g.:

   node attribute node1 set rack 1

If a server is moved to a different rack, simply update the value of its
attribute and it will use that rack's fencing configuration. Or if a
rack gets a new fencing device, you only have to update the fencing
configuration once rather than for every node in the rack.

The syntax accepts either '=' or ':' as the separator for the name/value
pair, so "rack:1:" would work in the CLI as well.

(Adapted for CLI syntax from announcement email)