medium: command: Disable fuzzy matcher for completion (#116)
There is negative interaction between the fuzzy matching when
executing commands (which allows using abbreviations like "cfg"
for "configure") and the bash completion. Instead of completing
"ve" to "verify", the completer would see them as the same word
and instead return the argument completions of "verify".

Disabling the fuzzy_get when completing should be enough to mostly
fix this, although it has the unfortunate side effect of breaking
argument completion when using abbreviations.