medium: scripts: Simplify SBD script (bsc#968076) (fate#318320)
Remove the Device parameter from the SBD script. The expectation
is that the device and all other parameters for SBD are configured
using /etc/sysconfig/sbd.

A possible improvement for this wizard would be to make it possible
to choose to create or configure an existing SBD device, which would
then involve updating /etc/sysconfig/sbd.

Another improvement would be to verify that sbd has been configured
on all nodes before applying the cluster configuration.