medium: history: Revert preference of messages over ha-log.txt (bsc#1031138)
The best fix to this problem would be to combine all available
log files for a node into a single stream. A temporary improvement
is to prefer ha-log.txt over messages, which should be a better
choice if hb_report selected the correct log file.

Partially reverts 2.3.1-3-g99f2056f.