medium: cibconfig: Sanitize CIB for patching (bsc#1127716)
It turns out that we DO need to sanitize the CIB before
generating the patch, or we'll generate a patch that wants
to remove all the things that the earlier sanitize call
wanted to remove. However, the part of sanitize_cib which
reordered some tags shouldn't be applied to the comparison
CIB, since doing that would mess up the reordering we did
when sanitizing the "new" CIB.

Confusing.

But this should work.