Dev: ui_resource: ask about ALL primitives when overriding attributes
When applying a maintenance attribute, the crmsh asked the user if he
wanted to delete the maintenance attribute for those primitives that had
a different value. It's more logical however to ask the user if he also
wants to delete the maintenance attribute from the primitives that have
the same value, because this is exactly the intention - to maintain the
group not each primitive separately. Even if the user would like to
customize each resource individually, he can answer 'No'.