Test: ui_resource: ask about ALL primitives when overriding attributes
The newly added test works as follows:

1) The primitive p5 gets the maintenance=true attribute.
2) The group g1 that contains the p5 primitive gets the
maintenance=true attribute. The test automatically aswers 'Yes' to the
question if the p5 maintenance attribute should be deleted. After all
only g1 has the maintenance attribute.

Previously, the user will only be asked to detele the maintenance attribute
from the p5 if it was set to false on the 1st step.