scripts: python3.8 unittest compatibility
When the unitetss are run with python 3.8 they break on changes in ast.
  File "/usr/lib/python3/dist-packages/crmsh/minieval.py", line 222, in _eval
    handler = self.nodes[type(node)]
  KeyError: <class '_ast.Constant'>

This is due to [1] which states "Class ast.Constant is now used for all
constants. Old classes ast.Num, ast.Str, ast.Bytes, ast.NameConstant and
ast.Ellipsis are still available, but they will be removed in future
Python releases.".
Even thou this is just the deprecation it already triggers an error
in the tests of crmsh.

[1]: https://docs.python.org/3/library/ast.html#ast.AST

Fixes #518

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>