Low: hb_report: Fix collecting of binary data (bsc#1166962)
When creating a resulting archive in a slave collector ('hb_report
__slave'), the code tried to decode output produced by the tar utility
as a UTF-8 string (crmutils.get_stdout_stderr() -> to_ascii()). This
corrupted any binary files in the archive, such as gzipped pengine
files.

The patch fixes the problem by obtaining the output from the tar utility
as a bytes object and writing it in the repr() form on stdout of the
slave collector. Similarly, the receiving side is updated to accept this
format.