Fix: bootstrap: check for missing fields in 'crm_node -l' output (bsc#1182131)
The bootstrap code expects that each line in the output from
'crm_node -l' is in the following form: <id> <nodename> <state>, with
all three fields always present. This is the usual situation but it is
in some cases possible that <nodename> and/or <state> can be missing.
For instance, a node can be configured in the nodelist in corosync.conf
but until the node has joined the cluster, 'crm_node -l' will report
only its id. This configuration can even happen as a result of
a previous incomplete 'crm cluster join' run.

Two places that parse 'crm_node -l' output in the bootstrap code are: in
setup_passwordless_with_other_nodes() when setting up passwordless SSH,
in join_cluster() when updating nodeids for IPv6. These functions can
fail with 'ValueError: not enough values to unpack (expected 3, got 1)'
or 'IndexError: list index out of range' when a nodename is missing in
the output.

The patch adds error handling for this situation. The code newly reports
a warning if a node state is incomplete and provides information what
operation is not possible to perform on the node. This allows bootstrap
of the new node to continue but indicates to the user that an update of
the old problematic node might be needed if it should rejoin the cluster
again in future.