Medium: ra: performance/usability improvement (avoid systemd)
Use --list-ocf-providers to check the crm_resource compatibility.
The --list-standards invokes systemd which takes quite some time.
Unfortunately, it doesn't seem like that will be fixed and, at any rate,
it's better to depend less on other system components.

This is no change in functionality, but a major improvement in
user experience. See #459 for more information.