Parametrize the log dir
By default all logs are stored in /var/lib/heartbeat/trace_ra
This patch adds a new resource attribute trace_dir to the cib.file
The trace_dir should come alongside the trace_ra,
whereas trace_ra signals that the resource is traced,
the trace_dir defines the path.
IMPORTANT: This patch changes the syntax of the $ crm resource trace
command. Now, the action should be prepanded with the -a key
and the interval with the -i key and the directory with the -d command.
Those keys are required because they are optional and can be umbiguous.