Dev: qdevice: Adjust qdevice reload policy
* Problem
When configuring sbd and qdevice together like crm cluster init -s <sbd>
--qnetd-hostname <qnetd> -y, corosync-qdevice.service will not start.

* Reason
According to the logic of evaluate_qdevice_quorum_effect, after
configured sbd and cluster started, there is an RA(external/sbd)
running, at this time, to use qdevice service, need to restart cluster
later.

* Solution
When adding qdevice and not in qdevice stage, this means that qdevice is
added from the beginning at init node. It's safe to restart the cluster,
regardless of whether there is RA running or not.