Dev: bootstrap: allow the cluster to operate with ssh session under non-root sudoer (jsc#PED-290)
Many servers do not permit root to login from ssh for enhanced security.
This change enable crmsh to operate the cluster with this setup.

The implementation is to do most of work as root, and su to a non-root user,
which is only used for ssh login, only when a command is needed to be executed
on a remote node. The remote command is executed in a `sudo /bin/sh` shell, so
it also has privilege.

This relies on crmsh being called with sudo, so that crmsh will be able to
determine which non-root user to su to when it needs to create a ssh
session. When crmsh is not called with sudo, it should work the same way
as before.