Dev: behave: use a standalone agent instead of crmsh.parallax to run commands
Background
==========

It is expected that crmsh uses the current user only unless another user
is specified explicitly. However, the code in functional tests needs operate
with different users to cover different use cases, requiring a complex
procedure to choose a user.

Changes
=======

A standalone agent is used to running command in functional test. It is
used only in function test, so the test code and production code is decoupled.