Fix: term: unset env `COLUMNS` and `ROWS` (bsc#1205925)
A bug in python stardard library, python/cpython#100516, make
environment variables, `COLUMNS` and `ROWS` to be added at fork. These
environment variables may make some of the commands called by crmsh to
truncate its output even if those commands are not writing to the
terminal.

These variables is set when bash enables its `checkwinsize` option. As
crmsh is able to read terminal size with curses, these variables are
unneeded and may be out of sync with actual terminal size.