Fix: report.utils: Fix the performance issue (bsc#1232821)
The function `extract_critical_log` can be very slow when processing
large log files. The main issue is the inefficiency of the regular
expression, which combines multiple wildcards (.*) with alternation (|),
leading to excessive backtracking. Additionally, the function reads the
entire file into memory, which is not optimal for large files.

To improve performance, it is better to use `grep` with the -F option to
search for fixed strings.