Dev: behave: Fix cross-network isolation issue
After the GitHub Actions environment was updated, the network isolation
between the two interfaces no longer functions as it did previously. As a
result, the test case "Setup cluster with crossed network" does not fail as
expected.

To address this, an iptables rule was added to simulate the isolated
network environment required for the test case.