Fix: log: missing LF after a progress bar (#1886)
In #1020 we right pad the progress bar with space characters so that it
will not mess up with log messages printed concurrently during the
progress bar runs. However, if the temrinal width changes after the
progress bar is printed, the message line following it will be shown in
the same line with it.

This commit fixes this by adding a final turn to print the progress bar
without right padding and with a LF character.

Fixes #1886.