Fix: sh: Avoid unnecessary sudo calls in ClusterShell SSH commands (bsc#1229416)
The `subprocess_run_without_input` method in `ClusterShell` was
unconditionally including `sudo -H -u ${user}` in its SSH command
construction, even when the target user is the same as the remote user.

This commit modifies the logic to conditionally include `sudo -H -u ${user}`
only when the target `user` is different from the `remote_user`. This
prevents unnecessary `sudo` calls, which is important in environments
where `sudo` might not be installed or configured, and improves the
robustness of SSH command execution.