Dev: ui_sbd: Refactor SBD crashdump set/purge flow to avoid inconsistent /etc/sysconfig/sbd
When setting SBD crashdump on a cluster running a non-stonith RA,
the command aborts and instructs the user to rerun with -F to
leverage maintenance mode. However, /etc/sysconfig/sbd could be
modified before the abort, leaving the file in an inconsistent state.

This commit refactors the set/purge workflow so /etc/sysconfig/sbd
is only updated after maintenance mode is leveraged,
and reorganize the logic to improve readability.