Dev: sbd: Adjust check/fix functions for watchdog timeout and msgwait timeout
For SBD_WATCHDOG_TIMEOUT for diskless SBD, and for watchdog timeout for
disk-based SDB, the expected value might be from profiles.yaml or the
default minimum value.

For msgwait timeout for disk-based SBD, the expected value should be the
maximum value from 2 times of the expected watchdog timeout and the runtime
msgwait timeout.